Use alloc version in place of precommit (#1465)
* use alloc version in place of precommit

* add timing log for finalize

* add upload changer timing logs

* use multiple txns

* add txn in mkdir

* remove lock in upload changer

* fix dest ref in mkdir

* add gorm logger

* go mod tidy

* add logger config for gorm

* use ptr for create

* change constraint name

* fix lookup hash constraint

* remove index from drop

* add created ref to cache

* use hash for lookup

* add cache logs

* fix change index

* remove deleted at clause

* add check for file ref

* add option for empty commit

* fix create marker in rollback handler

* add thumbnail suffix in write file

* add fix for move and copy

* add log

* use refLookupHash

* set num blocks

* add fix for destpath

* add max connection changes in config

* rmv get wm in file stats

* increment num_of_updates

+++ code/go/0chain.net/blobbercore/allocation/file_changer_upload.go
@@ -2,12 +2,16 @@ package allocation
 
 import (
 	"context"
+	"database/sql"
 	"encoding/json"
+	"math"
 	"path/filepath"
 	"strings"
+	"time"
 
 	"github.com/0chain/blobber/code/go/0chain.net/blobbercore/datastore"
 	"github.com/0chain/blobber/code/go/0chain.net/blobbercore/filestore"
+	"go.uber.org/zap"
 	"gorm.io/gorm"
 
 	"github.com/0chain/blobber/code/go/0chain.net/blobbercore/reference"
@@ -15,6 +19,7 @@ import (
 
 	"github.com/0chain/blobber/code/go/0chain.net/core/common"
 	"github.com/0chain/blobber/code/go/0chain.net/core/encryption"
+	"github.com/0chain/blobber/code/go/0chain.net/core/logging"
 )
 
 // swagger:model UploadFileChanger
@@ -25,12 +30,12 @@ type UploadFileChanger struct {
 
 // ApplyChange update references, and create a new FileRef
 func (nf *UploadFileChanger) applyChange(ctx context.Context,
-	ts common.Timestamp, _ map[string]string, collector reference.QueryCollector) error {
+	ts common.Timestamp, allocationVersion int64, collector reference.QueryCollector) error {
 
 	if nf.AllocationID == "" {
 		return common.NewError("invalid_parameter", "allocation_id is required")
 	}
-
+	now := time.Now()
 	parentPath := filepath.Dir(nf.Path)
 	nf.LookupHash = reference.GetReferenceLookup(nf.AllocationID, nf.Path)
 	newFile := &reference.Ref{
@@ -54,50 +59,62 @@ func (nf *UploadFileChanger) applyChange(ctx context.Context,
 		ChunkSize:               nf.ChunkSize,
 		CreatedAt:               ts,
 		UpdatedAt:               ts,
-		IsPrecommit:             true,
 		LookupHash:              nf.LookupHash,
 		DataHash:                nf.DataHash,
 		DataHashSignature:       nf.DataHashSignature,
 		PathLevel:               len(strings.Split(strings.TrimRight(nf.Path, "/"), "/")),
+		NumBlocks:               int64(math.Ceil(float64(nf.Size*1.0) / float64(nf.ChunkSize))),
 		FilestoreVersion:        filestore.VERSION,
+		AllocationVersion:       allocationVersion,
+		NumUpdates:              1,
 	}
-	newFile.FileMetaHash = encryption.Hash(newFile.GetFileMetaHashData())
-
+	newFile.FileMetaHash = encryption.FastHash(newFile.GetFileMetaHashData())
+	elapsedNewFile := time.Since(now)
 	// find if ref exists
 	var refResult struct {
-		ID   int64
-		Type string
+		ID         int64
+		Type       string
+		NumUpdates int64 `gorm:"column:num_of_updates" json:"num_of_updates"`
 	}
-	cachedRef := collector.GetFromCache(newFile.LookupHash)
-	if cachedRef != nil {
-		refResult.ID = cachedRef.ID
-		refResult.Type = cachedRef.Type
-	} else {
-		db := datastore.GetStore().GetTransaction(ctx)
-		err := db.Model(&reference.Ref{}).Select("id", "type").Where("lookup_hash = ?", newFile.LookupHash).Take(&refResult).Error
-		if err != nil && err != gorm.ErrRecordNotFound {
-			return err
-		}
+
+	err := datastore.GetStore().WithNewTransaction(func(ctx context.Context) error {
+		tx := datastore.GetStore().GetTransaction(ctx)
+		return tx.Model(&reference.Ref{}).Select("id", "type", "num_of_updates").Where("lookup_hash = ?", newFile.LookupHash).Take(&refResult).Error
+	}, &sql.TxOptions{
+		ReadOnly: true,
+	})
+	if err != nil && err != gorm.ErrRecordNotFound {
+		return err
 	}
+
 	if refResult.ID > 0 {
 		if !nf.CanUpdate {
 			return common.NewError("prohibited_allocation_file_options", "Cannot update data in this allocation.")
 		}
+		if refResult.Type != reference.FILE {
+			return common.NewError("invalid_reference_path", "Directory already exists with the same path")
+		}
 		deleteRecord := &reference.Ref{
-			ID: refResult.ID,
+			ID:         refResult.ID,
+			LookupHash: newFile.LookupHash,
+			Type:       refResult.Type,
 		}
 		collector.DeleteRefRecord(deleteRecord)
+		newFile.NumUpdates = refResult.NumUpdates + 1
 	}
+	elapsedNewFileRecord := time.Since(now) - elapsedNewFile
 	// get parent id
 	parent := filepath.Dir(nf.Path)
 	// create or get parent directory
-	parentRef, err := reference.Mkdir(ctx, nf.AllocationID, parent, ts, collector)
+	parentRef, err := reference.Mkdir(ctx, nf.AllocationID, parent, allocationVersion, ts, collector)
 	if err != nil {
 		return err
 	}
+	elapsedMkdir := time.Since(now) - elapsedNewFileRecord - elapsedNewFile
 	newFile.ParentID = &parentRef.ID
 	collector.CreateRefRecord(newFile)
-
+	elapsedCreateRefRecord := time.Since(now) - elapsedMkdir - elapsedNewFileRecord - elapsedNewFile
+	logging.Logger.Info("UploadFileChanger", zap.Duration("elapsedNewFile", elapsedNewFile), zap.Duration("elapsedNewFileRecord", elapsedNewFileRecord), zap.Duration("elapsedMkdir", elapsedMkdir), zap.Duration("elapsedCreateRefRecord", elapsedCreateRefRecord), zap.Duration("elapsedTotal", time.Since(now)))
 	return err
 }
